mlService/requirements.txt
flask
flask-cors
joblib
scikit-learn
google-generativeai
python-dotenv
